Lots of tweaks to get TLS validation in place.
Also starting to pave the way for http and dns validation. Currently the
default validation mode is TLS since the other two do not work yet. In
theory if you already had a validated account you wouldn't need to have
the DNS/HTTP challenges setup so it technically could work but it would
require leg work outside of the application.