Remove self signed cert and shutdown on startup
So previously we were doing this to make it so that if it failed the app
server would still be up and running but with a self signed certificate.
This is probably the wrong behavior.

So now instead if there is any error at all while requesting a
certificate the application will fail to start instead of being in a
state that is half working.